Treat waitset_data as a uintptr_t

+++ tenders/hvt/hvt.h
@@ -127,7 +127,7 @@ int hvt_vcpu_loop(struct hvt *hvt);
  * Register the file descriptor (fd) for use with HVT_HYPERCALL_POLL.
  * (waitset_data) must be set to the solo5_handle_t associated with (fd).
  */
-int hvt_core_register_pollfd(int fd, uint64_t waitset_data);
+int hvt_core_register_pollfd(int fd, uintptr_t waitset_data);
 
 /*
  * Register (fn) as the handler for hypercall (nr).
